Move search state to custom hook #634
