+++ src/main/resources/assets/react/search/useSearchState.ts
@@ -1,0 +1,262 @@
+import type {InterfaceField} from '/lib/explorer/types/Interface.d';
+import type {Hit} from './Hits';
+import type {
+	Profiling,
+	Synonyms
+} from './index.d';
+
+
+import * as gql from 'gql-query-builder';
+import * as React from 'react';
+
+
+type InterfaceName = string;
+type SearchString = string;
+
+export type SearchProps = {
+	//documentTypesAndFields ?:Record<string,Fields>
+	fields?: InterfaceField[]
+	firstColumnWidth?: 1|2|3|4|5|6|7|8|9|10|11|12|13|14|15
+	interfaceName?: InterfaceName
+	searchString?: SearchString
+}
+
+type SearchResult = {
+	count :number
+	hits :Array<Hit>
+	profiling :Array<Profiling>
+	locales :Array<string>
+	synonyms :Synonyms
+	total :number
+}
+
+
+export function useSearchState({
+	searchStringProp,
+	// Optional
+	fieldsProp = [],
+	interfaceNameProp = 'default',
+}: {
+	searchStringProp: SearchString
+	// Optional
+	fieldsProp?: InterfaceField[]
+	interfaceNameProp?: InterfaceName
+}) {
+	//console.debug('Search({fields:', fieldsProp,'})');
+	//console.debug('Search interfaceName', interfaceNameProp);
+
+	const [boolOnChange, setBoolOnChange] = React.useState(false);
+	//console.debug('Search boolOnChange', boolOnChange);
+
+	const [loading, setLoading] = React.useState(false);
+	//console.debug('Search loading', loading);
+
+	const [searchString, setSearchString] = React.useState(searchStringProp || '');
+	//console.debug('Search searchString', searchString);
+
+	//const [cache, setCache] = React.useState({} as Cache);
+	//console.debug('Search cache', cache);
+
+	const [result, setResult] = React.useState<SearchResult>({
+		count: 0,
+		hits: [],
+		profiling: [],
+		locales: [],
+		synonyms: [],
+		total: 0
+	});
+	//console.debug('Search result', result);
+
+	function search(ss :string) {
+		if(!ss) {
+			setResult({
+				count: 0,
+				hits: [],
+				locales: [],
+				profiling: [],
+				synonyms: [],
+				total: 0
+			});
+			return;
+		}
+		/*const cachedResult = getIn(
+			cache,
+			`${interfaceName}.${ss}`
+		) as SearchResult;
+		if (cachedResult) {
+			setResult(cachedResult);
+			return;
+		}*/
+		setLoading(true);
+		const uri = `./explorer/api/v1/interface/${interfaceNameProp}`;
+		//console.debug(uri);
+
+		const variables :Record<string, {
+			required ?:boolean
+			type ?:string
+			value :unknown
+		}> = {};
+		const fieldsHits :Array<
+			string
+			|Record<string,Array<string>>
+		> = [
+			'_collection',
+			'_documentType',
+		];
+		if (fieldsProp.length) {
+			fieldsHits.push('_highlight'); // GraphQLJson
+			variables['highlight'] = {
+				required: false,
+				type: 'InputTypeHighlight',
+				value: {
+					numberOfFragments: 1,
+					postTag: '</b>',
+					preTag: '<b>',
+					fields: fieldsProp.map(({
+						name
+					}) => ({
+						field: name
+					}))
+				}
+			}
+		}
+		fieldsHits.push('_json');
+		fieldsHits.push('_score');
+
+		const gqlQuery = gql.query({
+			operation: 'querySynonyms',
+			variables: {
+				//languages,
+				profiling: {
+					list: false,
+					required: false,
+					value: true // TODO Hardcode
+				},
+				searchString: {
+					list: false,
+					required: true,
+					value: searchString
+				}
+			},
+			fields: [
+				'languages',
+				{
+					profiling: [
+						'currentTimeMillis',
+						'label',
+						'operation'
+					]
+				},
+				{
+					synonyms: [
+						'_highlight',
+						'_score',
+						{
+							synonyms: [
+								'locale',
+								'synonym'
+							]
+						},
+						'thesaurusName'
+					]
+				},
+				{
+					operation: 'search',
+					fields: [
+						'count',
+						{
+							hits: fieldsHits
+						},
+						{
+							profiling: [
+								'currentTimeMillis',
+								'label',
+								'operation'
+							]
+						},
+						'total'
+					],
+					variables
+				}
+			]
+		}, null, {
+			operationName: 'InterfaceSearch'
+		});
+		//console.debug('Search() gqlQuery:', gqlQuery);
+		fetch(uri, {
+			method: 'POST',
+			headers: { 'Content-Type': 'application/json' },
+			body: JSON.stringify(gqlQuery)
+		})
+			.then(response => response.json())
+			.then(aResult => {
+				//console.debug('fetch aResult', aResult);
+				if (aResult && aResult.data && aResult.data.querySynonyms) {
+					const {
+						languages,
+						profiling,
+						search,
+						synonyms
+					} = aResult.data.querySynonyms;
+					const profilingArray = [];
+					const currentTimeMillisStart = profiling[0].currentTimeMillis;
+					//if(profiling) {
+					for (let i = 0; i < profiling.length - 1; i += 1) {
+						const durationMs = profiling[i + 1].currentTimeMillis - profiling[i].currentTimeMillis;
+						profiling[i + 1].durationMs = durationMs;
+						const durationSinceLocalStartMs = profiling[i + 1].currentTimeMillis - profiling[0].currentTimeMillis;
+						profiling[i + 1].durationSinceLocalStartMs = durationSinceLocalStartMs;
+						profiling[i + 1].durationSinceTotalStartMs = durationSinceLocalStartMs;
+						profilingArray.push(profiling[i + 1]);
+					}
+					//console.log('querySynonyms profiling', profiling);
+					//}
+					if (search) {
+						const {
+							profiling
+						} = search;
+						//if(profiling) {
+						for (let i = 0; i < profiling.length - 1; i += 1) {
+							const durationMs = profiling[i + 1].currentTimeMillis - profiling[i].currentTimeMillis;
+							profiling[i + 1].durationMs = durationMs;
+							const durationSinceLocalStartMs = profiling[i + 1].currentTimeMillis - profiling[0].currentTimeMillis;
+							profiling[i + 1].durationSinceLocalStartMs = durationSinceLocalStartMs;
+							const durationSinceTotalStartMs = profiling[i + 1].currentTimeMillis - currentTimeMillisStart;
+							profiling[i + 1].durationSinceTotalStartMs = durationSinceTotalStartMs;
+							profilingArray.push(profiling[i + 1]);
+						}
+						//console.log('search profiling', profiling);
+						//}
+						//console.log('search profilingArray', profilingArray);
+						search.locales = languages;
+						search.profiling = profilingArray;
+
+						//if (synonyms) {
+						search.synonyms = synonyms;
+						//}
+						/*setCache(prev => {
+							//console.debug('setCache prev', prev);
+							const deref = JSON.parse(JSON.stringify(prev));
+							if (synonyms) {
+								deref.synonyms = synonyms;
+							}
+							setIn(deref, `${interfaceName}.${ss}`, search);
+							return deref;
+						});*/
+						setResult(search);
+					}
+				}
+				setLoading(false);
+			}); // fetch
+	} // function search
+
+	React.useEffect(() => search(searchString), []);
+
+	return {
+		boolOnChange, setBoolOnChange,
+		loading, setLoading,
+		result, setResult,
+		search,
+		searchString, setSearchString,
+	}
+}
